requirements-core.txt
# Core ML and Data Science (Essential)
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Hyperparameter Optimization
optuna>=3.0.0

# Web Framework
Flask>=2.3.0
Flask-CORS>=4.0.0

# API and HTTP Requests
requests>=2.31.0

# Environment Variables
python-dotenv>=1.0.0

# Development and Testing
pytest>=7.4.0

# Performance and Memory
psutil>=5.9.0 